所有表主键转 big int 类型

+++ database/migrations/2019_05_07_200200_create_product_skus_table.php
@@ -14,12 +14,12 @@ class CreateProductSkusTable extends Migration
     public function up()
     {
         Schema::create('product_skus', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('title')->comment('SKU 名称');
             $table->string('description')->comment('SKU 描述');
             $table->decimal('price', 10, 2)->comment('SKU 价格');
             $table->unsignedInteger('stock')->comment('库存');
-            $table->unsignedInteger('product_id')->comment('所属商品 id');
+            $table->unsignedBigInteger('product_id')->comment('所属商品 id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
         });
